+++ experiment_server/_api.py
@@ -1,5 +1,24 @@
 from typing import Any, Dict, Union
-from experiment_server._main import GlobalState
+from experiment_server._process_config import process_config_file
+
+
+class GlobalState:
+    def __init__(self, config_file, participant_index):
+        self.config_file = config_file
+        self.change_participant_index(participant_index)
+
+    def change_participant_index(self, participant_index):
+        self._participant_index = participant_index
+        self._block_id = None
+        self.block = None
+        self.config = process_config_file(self.config_file, participant_index)
+
+    def set_block(self, block_id):
+        self._block_id = block_id
+        self.block = self.config[block_id]
+
+    def move_to_next_block(self):
+        self.set_block(self._block_id + 1)
 
 
 class Experiment: