+++ experiment_server/cli.py
@@ -2,7 +2,7 @@
 
 import click
 
-from experiment_server._main import _main
+from experiment_server._server import _server
 from experiment_server._process_config import verify_config
 
 @click.group()
@@ -17,7 +17,7 @@ def cli():
 @click.option("-p", "--port", default='5000')
 def run(participant_index, config_file, host, port):
     """Launch server with the `config-file` used to setup the configurations"""
-    _main(participant_index if participant_index > 0 else None, host, port, config_file)
+    _server(participant_index if participant_index > 0 else None, host, port, config_file)
 
 
 @cli.command()
@@ -25,10 +25,3 @@ def run(participant_index, config_file, host, port):
 def verify_config_file(config_file):
     """Verify if the config-file provided is valid"""
     verify_config(config_file)
-    
-# _main.add_command(test)
-# _main.add_command(server)
-    
-
-if __name__ == '__main__':  # pragma: no cover
-    _main()